updated threat model
- Minor rephrasing and white-space changes to make raw text nicer.
- Avoid using sigsum as "signed checksum" in text.  Not helpful.
- Removed paragraph about risk-averse attacker.  It is not needed to
make our points right now.  In a future revision, we should re-add this
and explain why it is interesting.  It would also be a good idea to then
cite the two papers that z4lem mentioned a while back, see archive.
- Clarified that we need a threshold of witnesses that follow the
cosigning protocol for security.  It is a start on addressing rohonk's
comment about which parties may (not) follow protocol and why.
- Emphasized that sigsum logging is only more course-grained than CT if
the data is actually lost.  Hence, more course-grained _in isolation_.
- Added links to slow-down and split-view attacks.