updated claimant model examples
I think the claimant model is most helpful for us to describe the
different use-cases of a sigsum log.  Let's focus on claimant models for
use-cases in this document, not claimant models for logs _as well_.

I kept R-B as an example because it is concrete, and fixed the long-due
updates that GeKo pointed out a while back about, e.g., "right data".