added a major refactor of design.md
The claimant model was mostly pulled from this document.  It is useful
to define use-cases of sigsum in a succinct way, but not helpful to tell
the reader about the concrete design that we have for a sigsum log.

(We still have a separate document that uses the claimant model.)

This refactor also tries to remove focus from use-cases that in general
are messy, and instead focus on the simple sigsum logging design that
has a very well-defined and thought-through usage-pattern.  The result
of this is that things should be a little bit more down-to-the-point.