keep summary session at its current location
I don't think it improves our design document by being moved.  We
already have a summary of properties in the introduction, and an easier
primer at the start of Section 3 that is strongly coupled to Figure 1.

Perhaps it is no longer necessary though.  When we wrote this we did not
have a summary of properties in introduction, or a relatively detailed
walk-through of the log's intended usage-pattern.

I'm fine with both keeping as is or deleting if it feels redundant.