documented the decided remove arbitrary bytes proposal
Refer to doc/proposals/2021-11-remove-arbitrary-bytes.md for details.

Since our proposal left the exact terminology undefined, this commit
took a stab at that.  The main idea was to keep referring to what we
have in a leaf and what is being signed as a _checksum_.  This ensures
that we are not undermining or stepping away from our core of "signed
checksums".  It seemed quite natural to refer to a checksum's preimage.