Update README and LICENCE
Focus on the content in this repository, not the project at large.

Further fix the licence to be more appropriate for documentation.  CC
BY-SA 4.0 is one of the available options when pre-printing academic
papers (https://arxiv.org/help/license) and likely a good fit for us.
We are already using this for brand assets and www.sigsum.org as well.