Revert "v1" label and multiple cosignature lines
As an "rc-1" interim for witness.md.  Expect major changes to the wire
API of witness.md "rc-2".  See the ongoing proposal for that in !46.

The decision to move this commit forward can be found here:

  https://git.glasklar.is/sigsum/project/documentation/-/blob/main/archive/2023-09-19--meeting-minutes.md?ref_type=heads#decisions

Strictly speaking multiple cosignature lines was not spelled-out, but it
should be a reasonable interpretation of dropping "v1" as an interim.