Minor improvements to www.so landing page
Slightly shorter initial sentence, "building block" is already coming as
a build-on in the next paragraph.

Also factor out the single use of "signed checksum", and instead
consistently just talk about transparent signatures on the landing page.
This is based on Filippo's feedback that the term is not explained well
in the getting started guide, and it is even less explained here.