www.so: Mark cosignature spec as v1
Also checkpoint and signed note are v1 now, but I think it's OK we're
not listing them here for now and instead just reference them from the
specs we do list.  Would make more sense to list when Sigsum also has a
checkpoint endpoint and tooling that uses it (instead of ASCII).