Order of doing things here does not matter
What's important is that we say this is a manual process, and we're
trying to say what parameters go into the selection (which we do).  But
whether we first look at witnesses or logs (for example) is irrelevant,
and in practise we will most likely be doing both things in parallel.